Edit setup to iterate through a list of config directories instead of having to individually supply the paths

+++ setup.sh
@@ -1,10 +1,6 @@
 #!/bin/bash
 
-declare -a directories=(
-  './vim'
-  # './zsh'
-  )
-
+# replace system config files with those in the dotfiles directory
 setup_config_files() {
   extension=$(echo $1 | cut -c 3-)
   date_and_time=$(date +'%m-%d-%y_%T')
@@ -16,22 +12,35 @@ setup_config_files() {
   # create current rc file path
   rc_file_path=~/."${extension}rc"
 
-  # backup and remove the current rc file
-  cp $rc_file_path "./$backup_dir/backup.$extension"
-  rm $rc_file_path
+  # backup and remove the current rc file, if one exists
+  if [[ -f $rc_file_path ]]
+  then
+    cp $rc_file_path "./$backup_dir/backup.$extension"
+    rm $rc_file_path
+  fi
 
-  # # create a symlink for the new rc to replace the old rc
-  ln -s "$PWD/$1/0-index.ignore.$extension" $rc_file_path
+  # create a symlink for the new rc to replace the old rc
+  dotfile_config_path="$PWD/$1/0-index.ignore.$extension"
+  ln -s $dotfile_config_path $rc_file_path
 
-  echo 'done!'
+  echo "Created symlink from $rc_file_path to $dotfile_config_path!"
 }
 
+# add the included fonts to the local fonts library
 installFonts() {
-  # add the included fonts to the local fonts library
   cp -r ./public/Fonts ~/Library
 }
 
-setup_config_files './vim'
-setup_config_files './zsh'
+# create an array of directories to pull config files from
+declare -a directories=(
+  './vim'
+  './zsh'
+)
+
+# run setup for each directory
+for directory in "${directories[@]}"; do
+  setup_config_files $directory
+done
+
 
 installFonts
